Deprecate Binding and move BindingImpl to Binding.kt
